requirements.txt
pipdeptree
absl-py==1.4.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
dm-control==1.0.10
dm-env==1.6
dm-tree==0.1.8
fasteners==0.18
fonttools==4.38.0
glfw==2.5.6
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.3
gym==0.21.0
h5py==3.8.0
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
kiwisolver==1.4.4
labmaze==1.0.6
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66

# works for sacae BUT NOT for vmail
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96

# works for vmail and sacae
# nvidia-cuda-nvrtc-cu11==11.8.89
# nvidia-cuda-runtime-cu11==11.8.89
# nvidia-cudnn-cu11==8.6.0.163

oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==4.22.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyOpenGL==3.1.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rl-plotter==2.4.0
robosuite==1.4.0
mujoco==2.3.2
# mujoco-py==2.1.2.14 (works)
# mujoco-py==2.0.2.5  (gives error)
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
tensorflow==2.12.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-probability==0.19.0
termcolor==2.2.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
pyquaternion
ipdb
wandb
# moviepy
colorlog
# mpi4py
# pybullet==1.9.5
plotly==2.7.0

# For SceneGrasp
fvcore
open3d
pygifsicle
pyrender    # use --no-deps option
dominate
shortuuid
boto3
zstandard
colour
pytorch_lightning==0.7.5